Make aggregated HAR types importable
